requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
SQLAlchemy==1.4.41
SQLAlchemy-serializer==1.4.1
typing-extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
PyJWT==1.7.1
Flask-RESTful==0.3.8